feat(readme)!: update tech events for 2024 and enchance visuals
BREAKING CHANGE: Added "Abacatinhos" logo to enhance visual appeal of the repository, aligning with our goal for a more engaging user experience.Archived 2023 tech events, ensuring historical
data is organized and accessible. Concurrently, the README has been updated with the list of 2024 tech events available as of this commit, reflecting our commitment to providing the most
current information. Introduced a month-wise summary in the README to improve navigation, enabling easier access to events by month and enhancing user interaction with the repository
content. Implemented general enhancements to the README, including improvements in the contribution guidelines, centralization of header elements for a cleaner visual hierarchy, and
strategic relocation of file paths to optimize repository structure.

closes issue #133